+++ CHANGELOG.md
@@ -32,3 +32,4 @@
 - [feat:upgrade api circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/1441)
 - [feat: support lossless config from console & support warmup.](https://github.com/Tencent/spring-cloud-tencent/pull/1452)
 - [feat:add admin http handler.](https://github.com/Tencent/spring-cloud-tencent/pull/1453)
+- [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1457)